Refactor away circular dependency
`Review` was depending on `DiffWindow` which was depending back on `Review`. `DiffWindow` didn't need _most_ of what was being provided in `Review` anyway.

Refactor out the common elements to remove the circular dependency and have `Review` and `DiffWindow` depend on the single common child component.
